refactor: split deck feature by responsibility
