+++ src/entities/deck/index.ts
@@ -3,4 +3,5 @@ export type { Category } from "./model/category";
 export { createDeck } from "./model/deck";
 export type { Deck, DeckEdit, DeckId, DeckNew } from "./model/deck";
 export { useDecks } from "./hooks/useDecks";
+export { useDeckMutations } from "./hooks/useDeckMutations";
 export { startDeckReads, stopDeckReads } from "./model/remoteReadStore";
