+++ src/features/deck-editor/components/templates/DeckFormTemplate.spec.tsx
@@ -9,7 +9,7 @@ import userEvent from "@testing-library/user-event";
 import { describe, expect, it, vi } from "vitest";
 import "@testing-library/jest-dom/vitest";
 
-import { DeckFormTemplate } from "@/features/deck/components/templates/DeckFormTemplate";
+import { DeckFormTemplate } from "./DeckFormTemplate";
 import { createDeck } from "@/test/factories";
 
 describe("DeckFormTemplate", () => {
