+++ src/features/deck-editor/components/templates/DeckFormTemplate.tsx
@@ -1,5 +1,5 @@
 /**
- * @file Composes the deck feature's complete Deck Form Template screen.
+ * @file Composes the deck editor feature's complete Deck Form Template screen.
  * Data and callbacks arrive through props, which keeps this presentation usable in both a live
  * container and Storybook.
  */
@@ -8,7 +8,7 @@ import type * as React from "react";
 import { AiOutlineArrowLeft } from "react-icons/ai";
 
 import { Layout } from "@/shared/ui/layout";
-import { DeckForm, type DeckFormProps } from "@/features/deck/components/DeckForm";
+import { DeckForm, type DeckFormProps } from "../DeckForm";
 
 export interface DeckFormTemplateProps {
   layout?: React.ComponentProps<typeof Layout>;
