+++ src/features/deck-editor/index.ts
@@ -1,0 +1,1 @@
+export { DeckFormContainer } from "./containers/DeckFormContainer";
