+++ src/features/deck-editor/lib/deckFormSchema.ts
@@ -1,5 +1,5 @@
 /**
- * @file Provides deck feature rules for Deck Form Schema.
+ * @file Provides deck editor rules for Deck Form Schema.
  * Keeping these calculations outside React makes their inputs, outputs, and edge cases easier to
  * understand and test.
  */
