--- src/features/deck/index.ts
@@ -1,5 +1,0 @@
-export { DeckStartForm, type DeckStartFormProps } from "./components/DeckStartForm";
-export { DeckFormContainer } from "./containers/DeckFormContainer";
-export { useDeckActions } from "./hooks/useDeckActions";
-export { useDeckFilterState } from "./hooks/useDeckFilterState";
-export { useDeckMutations } from "./hooks/useDeckMutations";
