+++ src/features/import/hooks/useDeckImport.ts
@@ -11,10 +11,9 @@ import { useCallback, useEffect, useRef, useState } from "react";
 
 import { documentMetadata as firestoreMetadata } from "@/adapters/firestore";
 import { createCard, selectCardsForDeck, useCards } from "@/entities/card";
-import { createDeck, useDecks } from "@/entities/deck";
+import { createDeck, useDeckMutations, useDecks } from "@/entities/deck";
 import { useSession } from "@/entities/session";
 import { useCardMutations } from "@/features/card/hooks/useCardMutations";
-import { useDeckMutations } from "@/features/deck/hooks/useDeckMutations";
 import type { DeckImportPreview, DeckImportResult, DeckImportRow } from "@/features/import/components/deckImportTypes";
 import { parseCsv } from "@/features/import/lib/cardCsv";
 import { buildDeckImportPlan, parseDeckImportCsv } from "@/features/import/lib/deckImportAnalysis";
