+++ src/features/study/components/templates/DeckStartTemplate.stories.tsx
@@ -7,8 +7,8 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { INITIAL_VIEWPORTS } from "@/storybook/storybookViewports";
-import { DeckStartForm, type DeckStartFormProps } from "@/features/deck/components/DeckStartForm";
-import { DeckStartTemplate as Template } from "@/features/study/components/templates/DeckStartTemplate";
+import { DeckStartForm, type DeckStartFormProps } from "../DeckStartForm";
+import { DeckStartTemplate as Template } from "./DeckStartTemplate";
 import * as fixture from "@/storybook/fixture";
 
 const deckStartForm: DeckStartFormProps = {
