+++ src/pages/deck-form/ui/DeckFormPage.tsx
@@ -1,3 +1,3 @@
-import { DeckFormContainer } from "@/features/deck";
+import { DeckFormContainer } from "@/features/deck-editor";
 
 export const DeckFormPage = () => <DeckFormContainer />;
