+++ src/pages/deck-list/ui/DeckListCard.tsx
@@ -1,5 +1,5 @@
 /**
- * @file Defines the deck feature's Deck Card presentation component.
+ * @file Defines the Deck List Page's Deck Card presentation component.
  * The component renders props and reports user intent through callbacks while data access stays
  * outside the view.
  */
